Create profile_scraper.py
in this script i attempted testing the accessibility of user profile pages on a website. The script uses the requests module to send a GET request to the URL of each user profile page, and checks the response for a successful status code (200). If the status code is 200, the script parses the JSON data in the response and extracts the user's name. It then prints the user's name to the console. If the status code is not 200, the script prints a message indicating that access to the user's profile page is restricted
The aim is to create a script that can be used to test for IDOR in a website